--- QUEERx/Home.js
@@ -1,11 +1,0 @@
-import { Text, View } from 'react-native';
-
-function HomeScreen() {
-    return (
-      <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
-        <Text>Home Screen</Text>
-      </View>
-    );
-}
-
-export default HomeScreen();
