--- QUEERx/constants.js
@@ -1,67 +1,0 @@
-// constants.js
-export const doctorsData = [
-    {
-        id: 4,
-        title: 'Dr. Kaur',
-        location: 'Palo Alto, CA',
-        description: 'Primary Care Physician, MD',
-        image: require('./assets/doctor4.png'),
-        filters:['distance', 'mi5', 'gender', 'woman'],
-        rating: 2
-    },
-    {
-        id: 1,
-        title: 'Dr. Jayakar',
-        location: 'Sunnyvale, CA',
-        description: 'Primary Care Physician, MD',
-        image: require('./assets/doctor1.png'),
-        filters:['distance', 'mi10', 'gender', 'woman', 'ratng', 'star4', 'star35', 'star3'],
-        rating: 4
-    },
-    {
-        id: 2,
-        title: 'Dr. Qureshi',
-        location: 'San Francisco, CA',
-        description: 'Primary Care Physician, MD',
-        image: require('./assets/doctor2.png'),
-        filters:['distance', 'mi25', 'gender', 'woman', 'ratng', 'star3'],
-        rating: 3
-    },
-    {
-        id: 3,
-        title: 'Dr. Posada',
-        location: 'Redwood City, CA',
-        description: 'Primary Care Physician, MD',
-        image: require('./assets/doctor3.png'),
-        filters:['distance', 'mi10', 'gender', 'man', 'ratng', 'star4', 'star35', 'star3'],
-        rating: 4
-    },
-    {
-        id: 6,
-        title: 'Dr. Singh',
-        location: 'Mountain View, CA',
-        description: 'Primary Care Physician, MD',
-        image: require('./assets/doctor6.png'),
-        filters:['distance', 'mi10', 'gender', 'woman'],
-        rating: 1
-    },
-    {
-        id: 5,
-        title: 'Dr. Suri',
-        location: 'Cupertino, CA',
-        description: 'Primary Care Physician, MD',
-        image: require('./assets/doctor5.png'),
-        filters:['distance', 'mi25', 'gender', 'man'],
-        rating: 1
-    },
-    {
-        id: 7,
-        title: 'Dr. Rajput',
-        location: 'Hayward, CA',
-        description: 'Primary Care Physician, MD',
-        image: require('./assets/doctor7.png'),
-        filters:['distance', 'mi50', 'gender', 'woman', 'ratng', 'star4', 'star35', 'star3'],
-        rating: 4
-    },
-  ];
-
