+++ QUEERx/package.json
@@ -9,36 +9,16 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@expo/webpack-config": "^19.0.0",
-    "@react-native-community/masked-view": "^0.1.11",
-    "@react-navigation/native": "^6.1.9",
-    "@react-navigation/native-stack": "^6.9.17",
-    "@react-navigation/stack": "^6.3.20",
-    "body-parser": "^1.20.2",
     "expo": "~49.0.15",
     "expo-status-bar": "~1.6.0",
-    "express": "^4.18.2",
     "react": "18.2.0",
-    "react-dom": "18.2.0",
     "react-native": "0.72.6",
-    "react-native-gesture-handler": "^2.14.0",
-    "react-native-paper": "^5.11.3",
-    "react-native-picker-select": "^9.0.0",
-    "react-native-reanimated": "^3.6.1",
-    "react-native-safe-area-context": "^4.7.4",
-    "react-native-screens": "^3.27.0",
-    "react-native-star-rating": "^1.1.0",
-    "react-native-vector-icons": "^10.0.2",
+    "react-dom": "18.2.0",
     "react-native-web": "~0.19.6",
-    "react-navigation": "^5.0.0",
-    "react-navigation-stack": "^2.10.4"
+    "@expo/webpack-config": "^19.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
   },
-  "private": true,
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": ""
+  "private": true
 }
