--- package.json
@@ -1,7 +1,0 @@
-{
-  "dependencies": {
-    "react-native-paper": "^5.11.1",
-    "react-native-safe-area-context": "^4.7.4",
-    "react-native-vector-icons": "^10.0.2"
-  }
-}
